Move platform.ts to src/file-system/controllers
