+++ src/App.tsx
@@ -7,6 +7,7 @@ import { ToastContainer } from "./components/ui/toast";
 import WelcomeScreen from "./components/window/welcome-screen";
 import { ZoomIndicator } from "./components/zoom-indicator";
 import { initializeThemeSystem } from "./extensions/themes/theme-initializer";
+import { isMac } from "./file-system/controllers/platform";
 import { useScroll } from "./hooks/use-scroll";
 import { useAppStore } from "./stores/app-store";
 import { useFileSystemStore } from "./stores/file-system/store";
@@ -18,7 +19,6 @@ import { useFontStore } from "./stores/font-store";
 import { useRecentFoldersStore } from "./stores/recent-folders-store";
 import { useZoomStore } from "./stores/zoom-store";
 import { cn } from "./utils/cn";
-import { isMac } from "./utils/platform";
 
 // Initialize theme system immediately when the module loads
 // This ensures themes are loaded before the settings store tries to apply them
