+++ src/components/editor/breadcrumb.tsx
@@ -3,11 +3,11 @@ import { useEffect, useRef, useState } from "react";
 import { createPortal } from "react-dom";
 import { EDITOR_CONSTANTS } from "../../constants/editor-constants";
 import FileIcon from "../../file-explorer/views/file.icon";
+import { readDirectory } from "../../file-system/controllers/platform";
 import { useBufferStore } from "../../stores/buffer-store";
 import { useFileSystemStore } from "../../stores/file-system/store";
 import { useUIState } from "../../stores/ui-state-store";
 import type { FileEntry } from "../../types/app";
-import { readDirectory } from "../../utils/platform";
 
 export default function Breadcrumb() {
   const buffers = useBufferStore.use.buffers();
